+++ server/src/controllers/todo.js
@@ -1,8 +1,10 @@
 import Todo from '../models/todo.js';
 
 export const getAllTodos = async (req, res) => {
+  const user_id = req.user._id;
+
   try {
-    const todos = await Todo.find();
+    const todos = await Todo.find({user_id}).sort({createdAt: -1});
     res.status(200).json(todos);
   } catch (err) {
     res.status(500).json({ message: err.message });
@@ -13,11 +15,12 @@ export const createTodo = async (req, res) => {
   const todo = new Todo({
     title: req.body.title,
     description: req.body.description,
+    user_id: req.user._id,
   });
 
   try {
     const newTodo = await todo.save();
-    const updatedTodos = await Todo.find();
+    const updatedTodos = await Todo.find({user_id:req.user._id});
     res.status(201).json({message: "new todo added successfully", data:{newTodo:newTodo, todos: updatedTodos}});
   } catch (err) {
     res.status(400).json({ message: err.message});
@@ -25,9 +28,10 @@ export const createTodo = async (req, res) => {
 };
 
 export const deleteAllTodos = async (req,res) => {
+  const user_id = req.user._id;
     try {
-        await Todo.deleteMany({});
-        const todos = await Todo.find();
+        await Todo.deleteMany({user_id});
+        const todos = await Todo.find({user_id});
         res.status(200).json({message:"All todos deleted successfully", data:{todos}});
     } catch (error) {
         res.status(500).json({message: error.message})
@@ -36,12 +40,13 @@ export const deleteAllTodos = async (req,res) => {
 
 export const deleteTodoById = async (req,res) => {
     const {id} = req.params;
+    const user_id = req.user._id;
 
     try {
         const deletedTodo = await Todo.findByIdAndDelete(id);
         if(!deletedTodo) return res.status(404).json({message:"todo not found"});
         
-        const updatedTodos = await Todo.find();
+        const updatedTodos = await Todo.find({user_id});
 
         res.status(200).json({message: `todo with id ${id} deleted successfully`, data: {deletedTodo,todos:updatedTodos }});
     } catch (error) {
@@ -52,12 +57,13 @@ export const deleteTodoById = async (req,res) => {
 export const updateTodo = async (req,res) => {
     const {id} = req.params;
     const {title,description, completed} = req.body;
+    const user_id = req.user._id;
 
     try {
         const updatedTodo = await Todo.findByIdAndUpdate(id,{title,description,completed}, {new: true});
         if(!updateTodo) return res.status(404).json({message:"todo not found"});
 
-        const todos = await Todo.find();
+        const todos = await Todo.find({user_id});
         res.status(200).json({message: "todo updated successfully", data: {updatedTodo,todos}})
     } catch (error) {
